Update interview calendar page
